% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-Seurat_Object.R
\name{CreateSeurat_Pango}
\alias{CreateSeurat_Pango}
\title{Origi_Data_Type H5Seurat H5ad H510X}
\usage{
CreateSeurat_Pango(
  Origi_Data_Path,
  Spatial_Matrix_Path = NULL,
  Origi_Data_Type = "H5Seurat"
)
}
\description{
Origi_Data_Type H5Seurat H5ad H510X
}
