% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-Shiny_PangoICH.R
\name{.gene_distribution_map_shiny_Server}
\alias{.gene_distribution_map_shiny_Server}
\title{shiny server}
\usage{
.gene_distribution_map_shiny_Server(H5_Database_Path, Spatial_Matrix_Path)
}
\description{
shiny server
}
